加上時間判斷

+++ src/3.1/IHttpClientFactory/QueryMaskSqliteSample/Controllers/MaskController.cs
@@ -50,27 +50,34 @@ namespace QueryMaskSample.Controllers
 
             async Task MaskDataUpdate()
             {
-                var nextTime = _maskContext.MedicalMasks.First().UpdateTime.AddSeconds(600);
+                var updateTime = _maskContext.MedicalMasks.First().UpdateTime;
+                var nextTime = updateTime.AddSeconds(600);
                 var nowTime = DateTime.Now;
                 if (nextTime < nowTime)
                 {
-                    _maskContext.RemoveRange(_maskContext.MedicalMasks);
                     var maskInfos = await _maskService.GetMaskInfo();
-                    foreach (var maskInfo in maskInfos)
+                    var sourceTime = DateTime.Parse(maskInfos.First().來源資料時間);
+                    // 0 is same as value
+                    // Ref: https://docs.microsoft.com/en-us/dotnet/api/system.datetime.compareto?view=netcore-3.1#remarks
+                    if (sourceTime.CompareTo(updateTime) != 0)
                     {
-                        _maskContext.Add(new MedicalMask()
+                        _maskContext.RemoveRange(_maskContext.MedicalMasks);
+                        foreach (var maskInfo in maskInfos)
                         {
-                            Code = maskInfo.醫事機構代碼,
-                            Name = maskInfo.醫事機構名稱,
-                            Address = maskInfo.醫事機構地址,
-                            Tel = maskInfo.醫事機構電話,
-                            AdultNum = int.Parse(maskInfo.成人口罩剩餘數),
-                            KidNum = int.Parse(maskInfo.兒童口罩剩餘數),
-                            UpdateTime = DateTime.Parse(maskInfo.來源資料時間)
-                        });
-                    }
+                            _maskContext.Add(new MedicalMask()
+                            {
+                                Code = maskInfo.醫事機構代碼,
+                                Name = maskInfo.醫事機構名稱,
+                                Address = maskInfo.醫事機構地址,
+                                Tel = maskInfo.醫事機構電話,
+                                AdultNum = int.Parse(maskInfo.成人口罩剩餘數),
+                                KidNum = int.Parse(maskInfo.兒童口罩剩餘數),
+                                UpdateTime = DateTime.Parse(maskInfo.來源資料時間)
+                            });
+                        }
 
-                    await _maskContext.SaveChangesAsync();
+                        await _maskContext.SaveChangesAsync();
+                    }
                 }
             }
 
@@ -81,7 +88,6 @@ namespace QueryMaskSample.Controllers
                 await MaskDataUpdate();
 
                 return Ok(_maskContext.MedicalMasks.AsNoTracking());
-
             }
             catch (HttpRequestException ex)
             {
